* Restored migrations

* Uppdate appsettings.json and program

* Update Index page

* Update Address and address.js

* Merge dev-address into dev (#37)

* Update Address entity

* Update Address

* Renamed AddressLine in PlaceName in address entity

* Add migrations

* Update AddressCommand

* Update Index

* Update Product pages and validators

* Update Address and Controller

* Update RestaurantDashboard

* UpdateRestaurantList

* Update scripts in RestaurantList

* Update RestaurantList

* Update RestaurantList

* Update RestaurantList

* Update RestaurantDetail razor page

* Update RestaurantDetail

* Add localization and culture in RestaurantDetail

* Fix issues displaying currency and prices

* Update quantity in Commands

* Update supported cultures

* Update RestaurantDetail

* Update Basket razor page

* Update NuGet packages version

* Update README.md

* Revert "Update NuGet packages version"

This reverts commit 4eb562fe87eb437cc3d18d17e1a9c2cead285f60.

* Fix issue in UpdateBasketCommand

* Fix counting quantity in UpdateBasketCommand

* Fix quantity issue in RestaurantDetail razor page

* Update Basket razor page

* Fix price issue in UpdateBasketItemCommand and DeleteBasketItemCommand

+++ src/Application/OrderItems/Commands/DeleteOrderItem/DeleteOrderItemCommand.cs
@@ -4,6 +4,7 @@ using DeliveryWebApp.Domain.Entities;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
+using DeliveryWebApp.Application.Products.Commands.UpdateProducts;
 
 namespace DeliveryWebApp.Application.OrderItems.Commands.DeleteOrderItem
 {
@@ -15,10 +16,12 @@ namespace DeliveryWebApp.Application.OrderItems.Commands.DeleteOrderItem
     public class DeleteOrderItemCommandHandler : IRequestHandler<DeleteOrderItemCommand, OrderItem>
     {
         private readonly IApplicationDbContext _context;
+        private readonly IMediator _mediator;
 
-        public DeleteOrderItemCommandHandler(IApplicationDbContext context)
+        public DeleteOrderItemCommandHandler(IApplicationDbContext context, IMediator mediator)
         {
             _context = context;
+            _mediator = mediator;
         }
 
         public async Task<OrderItem> Handle(DeleteOrderItemCommand request, CancellationToken cancellationToken)
@@ -29,6 +32,22 @@ namespace DeliveryWebApp.Application.OrderItems.Commands.DeleteOrderItem
             {
                 throw new NotFoundException(nameof(Order), request.Id);
             }
+
+            var product = await _context.Products.FindAsync(entity.ProductId);
+
+            if (product == null)
+            {
+                throw new NotFoundException(nameof(Product), entity.ProductId);
+            }
+
+            // update product quantity
+            var newQuantity = product.Quantity + entity.Quantity;
+
+            await _mediator.Send(new UpdateProductCommand
+            {
+                Id = product.Id,
+                Quantity = newQuantity
+            }, cancellationToken);
 
             _context.OrderItems.Remove(entity);
             await _context.SaveChangesAsync(cancellationToken);
